+++ src/Git_Updater/Shim.php
@@ -35,6 +35,7 @@ if ( ! function_exists( 'move_dir' ) ) {
 	 *
 	 * @param string $from        Source directory.
 	 * @param string $to          Destination directory.
+	 *
 	 * @return true|WP_Error True on success, WP_Error on failure.
 	 */
 	function move_dir( $from, $to ) {
@@ -45,19 +46,19 @@ if ( ! function_exists( 'move_dir' ) ) {
 		/**
 		 * Fires before move_dir().
 		 *
-		 * @since 6.1.0
+		 * @since 6.2.0
 		 */
 		do_action( 'pre_move_dir' );
 
 		if ( 'direct' === $wp_filesystem->method ) {
-			$wp_filesystem->rmdir( $to );
-
-			$result = @rename( $from, $to );
-		}
-
-		// Non-direct filesystems use some version of rename without a fallback.
-		if ( 'direct' !== $wp_filesystem->method ) {
+			if ( $wp_filesystem->rmdir( $to ) ) {
+				$result = @rename( $from, $to );
+				wp_opcache_invalidate_directory( $to );
+			}
+		} else {
+			// Non-direct filesystems use some version of rename without a fallback.
 			$result = $wp_filesystem->move( $from, $to );
+			wp_opcache_invalidate_directory( $to );
 		}
 
 		if ( ! $result ) {
@@ -78,11 +79,68 @@ if ( ! function_exists( 'move_dir' ) ) {
 		/**
 		 * Fires after move_dir().
 		 *
-		 * @since 6.1.0
+		 * @since 6.2.0
 		 */
 		do_action( 'post_move_dir' );
 
 		return $result;
+	}
+}
+
+if ( ! function_exists( 'wp_opcache_invalidate_directory' ) ) {
+	/**
+	 * Invalidate OPcache of directory of files.
+	 *
+	 * @since 6.2.0
+	 *
+	 * @global WP_Filesystem_Base $wp_filesystem WordPress filesystem subclass.
+	 *
+	 * @param string $dir The path to invalidate.
+	 *
+	 * @return void
+	 */
+	function wp_opcache_invalidate_directory( $dir ) {
+		global $wp_filesystem;
+
+		if ( ! is_string( $dir ) || '' === trim( $dir ) ) {
+			$error_message = sprintf(
+			/* translators: %s: The '$dir' argument. */
+				__( 'The %s argument must be a non-empty string.', 'git-updater' ),
+				'<code>$dir</code>'
+			);
+			// phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_trigger_error
+			trigger_error( esc_html( $error_message ) );
+			return;
+		}
+
+		$dirlist = $wp_filesystem->dirlist( $dir, false, true );
+
+		if ( empty( $dirlist ) ) {
+			return;
+		}
+
+		/*
+		 * Recursively invalidate opcache of nested files.
+		 *
+		 * @param array  $dirlist Array of file/directory information from WP_Filesystem_Base::dirlist().
+		 * @param string $path    Path to directory.
+		 */
+		$invalidate_directory = function( $dirlist, $path ) use ( &$invalidate_directory ) {
+			$path = trailingslashit( $path );
+
+			foreach ( $dirlist as $name => $details ) {
+				if ( 'f' === $details['type'] ) {
+					wp_opcache_invalidate( $path . $name, true );
+					continue;
+				}
+
+				if ( is_array( $details['files'] ) && ! empty( $details['files'] ) ) {
+					$invalidate_directory( $details['files'], $path . $name );
+				}
+			}
+		};
+
+		$invalidate_directory( $dirlist, $dir );
 	}
 }