+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* update `Shim` for `move_dir()` and `wp_opcache_invalidate_directory()`
 
 #### 12.0.2 / 2023-01-12
 * PHP 8.1 compatibility fix, thanks @chesio